replace igraph with own adjacency matrix
Replace igraph with a homebrewed adjacency matrix type, as igraph
was being sorely underutilized (and an adjacency matrix is not hard
to implement)

As an unexpected bonus, the new class appears to be more performant
than the old one (but not too much!).

The old implementation is still there under the name
IgraphUndirectedGraph.  I'm keeping it around as it might be useful
to have an *adapter* for igraph objects (after all, you can do some
neat stuff in igraph).  That said, I'm not sure how inclined
I am to continue updating it...