Optimization: remove filaments
The algorithm runtime seemed to vary wildly based on the initial
rotation (9s to 35s for a 400x60 cell bridge).  One possible cause
I suspected is that some rotations may result in the algorithm repeatedly
retracing long filaments. (chains of edges that only connect to the graph
at one end)

This fix, which removes such filaments when they are created, consistently
performs 7.5s on the above-mentioned graph.