niche optimization: no linear search through path
after playing around with some graphs with *really* large cycles,
I figured I might as well put this in. (that and I think it reads
better)

That said, you need a graph with an extremely large cycle for it to
matter.  Around 10000 (`nx.cycle_graph(10000)`) is where it starts
becoming noticeable. Around 100000 is where it becomes life-saving.