abstract out method of node deletion
This will allow for trying out different "representations" of a
defect in the graph.

Right now, the only available method is to delete the node. An
example of another possible method would be to increase the
resistances of any attached edges by some factor.

(making this change caused a bunch of questions to arise with regards
 to how to reduce coupling, many of which I'm not sure how to address.
 At the very least, node_selection now only takes the initial graph,
 because there's no longer any guarantees on how the initial and current
 graphs differ.)