Split up compute_currents to aid in profiling
Things are looking grim for my homebrewed, incomplete graph classes;
not only are they a timesink -- not only is their presence bound to
confuse anyone new to the code base -- but their methods are also where
the program spends the vast majority of its time... far more so than the
sparse matrix solver.

At this point, it's more or less indefensible to keep using them.