MeshCurrentSolver
Okay.  The previous refactorings were alright.  This one is...
fairly less so, but something had to be done to gain back all the
speed that was lost on the previous refactorings.

It also aids somewhat in adding support for non-planar graphs, which
will be needed *very soon* (one word: MoS2)

That said, I wasn't sure how best to deal with the problem of
cache invalidation.  I tried to implement a simple solution...
...but needless to say, "tried" is the operative word.