cleaner graph validation func from mos2.py
When I was writing mos2.py, I had to stop and think quite a bit
reading hex-bridge's graph validation function.

The rewritten function reads much clearer, and gives a less
misleading error message. (the old one provided a list of
problems, but the list it provided was not necessarily complete!)

(It's also a fair bit slower, seemingly due to shenanigans in
 networkx's edges_iter... eh.)