build_cyclebasis
Implemented the cyclebasis "generation" algorithm that I probably
should've went with all along: manual.

More specifically, we provide a function which accepts a list of
'desirable' cycles and a reliably-generated 'fallback' cycle basis
(such as the one generated by networkx).

By maintaining a bit matrix of cycle edges in reduced row echelon
form, it gathers a linearly-independent subset of cycles from the
desirable set and completes it by adding any additional linearly-
independent cycles it identifies in the fallback.

For regular structures, it shouldn't be too difficult to generate
an almost comprehensive list of desirable, short-length cycles,
so that ultimately only a small number of the suboptimal 'fallback'
cycles are required.