Prepare for new circuit format
For too long, resistances.py has accepted as input a pickled
networkx graph containing attributes that describe a circuit
together with attributes that set parameters for the defects
trial, and possibly attributes that define a planar embedding.

Not only does using pickled data as an input file sound like
a *terrible idea* for *every possible reason ever*, the "one
file fits all" model also proves troublesome when trying to
write other scripts which use the files.

So we're going to break things up:

    my-graph.circuit
    my-graph.cycles
    my-graph.defects.conf
    my-graph.planar.xy
    my-graph.visual.xy

etc.

Which is a whole lot more sensible than the alternative way
I was considering to deal with this up until now...