add fixed_order, accomodate stateful selectors
The ``fixed_order`` selection mode replays a fixed order of
selections.

It was not possible to implement this under the already existing
model for selection modes as it required mutable state.  The newly
introduced types separate the stateless part (a "selection mode")
from the stateful part (a "selector") which exists during a trial.

This also allowed me to remove some extraneous state that
resistances.py was keeping track of, as the selectors can now
keep track of it on their own.