tear apart makefig.py
https://github.com/ExpHP/data

a magical land where it is *perfectly okay* for a plot to have
"hacky" things like hardcoded x-y ticks and hand-placed labels
based on features specific to a certain data set.

It's time for me to finally admit that it is simply *not possible*
to take the principles of writing maintainable code and apply them
when working with data.  (not without a profoundly negative impact
on your sanity, at least)