Small big things
Output functions resembling the logging interface, and a ``die``
function (which does that and aborts), which have become staples
in my more recent scripts (especially the latter, as otherwise
it can be easy to forget to abort after an error message).

Also take advantage of the ``type`` argument in argparse to take
care of the simpler parts of validation.