give `node_deletion` finer control
Up until now, `CurrentMeshSolver` implemented functions that did
pretty much exactly what the node deletion methods needed.

The issue is that this doesn't work so well once you have e.g.
defect representations that extend to adjacent nodes.  In this
case, the nodes that are "selected" and the nodes that are "removed"
are not necessarily the same any more, and before I can comfortably
allow the code to make a distinction between the two, node deletion
methods need to be aware of precisely which nodes they are deleting.

So now CurrentMeshSolver provides a simpler interface (modifying
single edges/nodes) and a node deletion method may end up calling
multiple methods.

...this also requires CurrentMeshSolver to provide an API for
inspecting the state of the graph, which is extremely unfortunate
for me as I know I'll end up wasting far too much time fretting
over silly details. :(