Get `defect-trial` running with new radius option
"Running", but not reasonably.

The whole "selected" vs "deleted" vs "what the heck are we removing
from `choices` this round?" thing started to make my head hurt, so
I went ahead and ignored that bit for now.  (when `annihilation` tries
to delete a vertex that's already removed, it just does nothing now...
which isn't quite what I want)

Unsurprisingly, I ran into another issue I didn't anticipate:
The way I used to protect the battery was by removing it from the
set of `choices` in the trial.  This doesn't work for radius >1
(nor did it ever work quite as intended, as the "assign resistance"
 mode was capable of giving resistance to the zero-resistance edges
 that don't actually belong to the system)

I adressed this for now with a hack that only works for very small
radii (up to 3).