modify `annihilation` behavior; add `Deleter`
Modify `annihilation` mode to be able to delete former neighbors
of a deleted node when `radius>1` (this way, it continues to act
as the limiting case for the "multiply resistance" mode when the
factor goes to infinity).

This requires it to remember the initial graph state, which pushed
me to create a `Deleter` class with internal state, which parallels
the `Selector` class for `SelectionMode`. (although since `Deleter`
only has one callback, a closure probably would've sufficed).

Some tough decisions still need to be made for the defect trial...