Get trials running again; redefine "defect"
Made a decision on how to define "defects" now that there is no
longer an obvious 1-1-1 correspondence between selected nodes,
defects, and nodes which are removed from `choices`.

The number of defects is now defined as
`len(initial_choices)-len(choices)`, and the selected nodes are
now called "event centers."  There is still 1 selected node per
substep... however, the number of defects per substep may vary.
There is an artificial restraint in place that this number must
be >= 1, as I would consider any other situation to be in error.
(e.g. a deleter returned neighbors of the center but not the
 center itself)