Improve handling of the `cannot_touch` param
I was wrong before in thinking that `cannot_touch` would allow
usage of radii up to 3. It only worked up to radius 2; anything
larger, and selecting one of the nodes at the edge would end up
deleting EVERY node that touches the battery, instantly disconnecting
the circuit.

So that's fixed.