Add missing `__init__.py` so tests are found
Hmm, that's unsettling... It's pretty easy to miss the fact
that a test isn't running! (and it's pretty easy to forget
to add an `__init__.py`!)

I guess it's not an issue so long as you follow the principle of
"no code without a failing test first"... but even then, the issue
may arise if you move existing modules (along with their already-
written tests).  Not sure how to prevent that.