Tests for `node_selection.py`
These are some particularly ugly tests as we're testing
randomness.

The current version of the test module uses code generation as
an intermediate step to let the test implementor manually review
some of the parameters that will be used by the test.

This was chosen as an alternative to implementing some kind of
heuristic for detecting poor parameters.