TrialRunner class
Up until now, the ``main`` method in ``resistances.py`` has produced
most of the variables for the trial runner directly from the command
line args.  This is one of the primary reasons no tests have been
written yet for the trial runner (as they'd either have to be CLI-based,
or they'd have to duplicate a bunch of code from ``main()`` which
unfortunately depends on a lot of unstable internal details)

TrialRunner does some of this work and handles setting up default
values for things, with the hope that tests for the trial runner will
at some point be feasible.

It also may or may not make it easier to eventually migrate some of
the parameters out of CLI arguments (which are really getting messy)
and into the config file.