Decouple cbproviders and cbupdaters
It kind of made sense back when it was implemented (as the planar
update method was only valid for planar cbs... but it isn't even
used now!), but not so much any more.  They're completely orthogonal
features to each other.

Now it might be possible to get rid of the ``from_existing``
provider...