Add tests for trial runner
This has been a long time coming, from the creation of the `Selector`
class (to allow a "`fixed_order`" selection method) to the recent
refactoring of the trial runner. We finally have everything we need
to make this feasible.

The purpose of these tests are to alert the developer when an update
to the codebase unintentionally causes the defect trial to compute
different currents.  They are designed in such a way that if a change
to the output *is* deliberate, it is easy to begin accepting the
the new results. (just rename a file... after making sure the new
results are sensible, of course!)